6º Semestre/Processamento de Imagens/pid-trabalho-2-2021/requirements.txt
python-mnist>=0.7
opencv-python>=4.5.4
numpy>=1.21.3
matplotlib>=3.4.3
mahotas>=1.4.12
tensorflow>=2.6.1
keras-datasets==0.1.1a1
scikit-learn>=1.0.1
sklearn